% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundInformation.R
\name{backgroundInformation}
\alias{backgroundInformation}
\title{Generate Random Background Points from Raster Extent}
\usage{
backgroundInformation(rasterLayers, n = 10000)
}
\description{
Samples a specified number of random background points from the non-NA cells
within the spatial extent defined by the input raster layers. The first layer
of the input SpatRaster is used as the template for identifying valid sampling locations.
}
