% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_multiple_layers.R
\name{download_multiple_layers}
\alias{download_multiple_layers}
\title{Download Multiple Bio-Oracle Layers}
\usage{
download_multiple_layers(
  variables,
  experiment,
  decade,
  latitude = NULL,
  longitude = NULL,
  exportRaster = TRUE,
  directory = NULL
)
}
\description{
Downloads specified Bio-Oracle layers based on combinations of variables,
predictors, experiments, realms, and decades. Allows spatial subsetting
and optionally returns a combined multi-layer SpatRaster.
}
