% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModelData.R
\name{generateModelData}
\alias{generateModelData}
\title{Generate Training, Validation, and Test Datasets for SDM}
\usage{
generateModelData(
  records,
  environmentalLayers,
  method = "random",
  proportionTrain = 0.7,
  proportionTest = 0.15,
  proportionVal = 0.15,
  bootstrapRounds = 6,
  paMinimum = 100,
  paRatio = 1,
  sacDist = NULL
)
}
\description{
splitting data into train/validation/test sets using either random folds or
spatial blocks, optionally creating multiple data replicates by re-sampling
absences (useful when presence data is scarce), extracting environmental
predictor values, and generating plots of the data splits.
}
