% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlocks.R
\name{getBlocks}
\alias{getBlocks}
\title{Assign Spatial Block Membership using K-Means Clustering}
\usage{
getBlocks(records, kFolds)
}
\description{
Assign Spatial Block Membership using K-Means Clustering
}
