% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerformance.R
\name{modelPerformance}
\alias{modelPerformance}
\title{Calculate Model Performance Metrics}
\usage{
modelPerformance(observed, predicted, index)
}
\description{
Evaluates model predictions against observed presence/absence values using
various threshold-dependent and independent metrics. It utilizes functions
from the 'sdm' and 'ecospat' packages and calculates AIC and R-squared.
Allows selection of metrics based on a specific threshold criterion.
}
