% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoAbsences.R
\name{pseudoAbsences}
\alias{pseudoAbsences}
\title{Generate Pseudo-Absence Points Excluding Buffer Around Presences}
\usage{
pseudoAbsences(rasterLayers, presences, n = 1000)
}
\description{
Creates a SpatVector of pseudo-absence points by randomly sampling locations
from the non-NA cells of a template raster layer, while excluding areas
within a specified buffer distance around known presence points.
}
