% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModelData.R
\name{generateModelData}
\alias{generateModelData}
\title{Generate Training, Validation, and Test Datasets for SDM}
\usage{
generateModelData(
  records,
  envConditions,
  method = "random",
  kFolds = NULL,
  proportionTrain = 0.7,
  proportionTest = 0.15,
  proportionVal = 0.15,
  paMinimum = 100,
  paRounds = 10,
  paRatio = 1
)
}
\description{
Prepares data for model training and evaluation by cleaning records,
splitting data into train/validation/test sets using either random folds or
spatial blocks, optionally creating multiple data replicates by re-sampling
absences (useful when presence data is scarce), extracting environmental
predictor values, and generating plots of the data splits.
}
