% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalDataGbif.R
\name{getExternalDataGbif}
\alias{getExternalDataGbif}
\title{Download Occurrence Data from GBIF}
\usage{
getExternalDataGbif(taxa, getCitation)
}
\description{
Fetches species occurrence records from the Global Biodiversity Information
Facility (GBIF). It handles potentially large datasets by downloading
records in chunks if the total count exceeds a threshold. Optionally retrieves
dataset citation information.
}
