% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExternalDataObis.R
\name{getExternalDataObis}
\alias{getExternalDataObis}
\title{Download Occurrence Data from OBIS}
\usage{
getExternalDataObis(taxa, getCitation)
}
\description{
Fetches species occurrence records from the Ocean Biodiversity Information
System (OBIS) using the \code{robis} package. Optionally retrieves dataset
citation information via the OBIS API.
}
